+++ src/main/java/com/library/project/web/repository/ILibroRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 import com.library.project.web.models.Libro;
 
-public interface ILibroRepository extends JpaRepository<Libro, Integer>{
+public interface ILibroRepository extends JpaRepository<Libro, Long>{
 
 }
